Save uploaded TV episodes
Implement the Episode.save method properly, chunking the HttpData and
writing it to the appropriate file.

Add config for max file sizes and the default chunk size used in
uploading

Add a config util to parse simple file size strings like 100K and 35M
into Int numbers of bytes. Might want to consider making that a Long
though.

Tweak the views a bit