Eliminate fullpath in favour of file
It makes much more sense to just have a file reference, rather than a
String path: it's usually just turned into a File straight away.

Replace cases where fullpath was being used as the start of a new File
path, with x.file.getCanonicalPath. That strips out any ..s and whatnot.

Add an exists method to the Media trait, as it's useful for both
Programme and Episode, as a shortcut to x.file.exists