Add basic authentication to all endpoints
Use a BasicAuth, configured under auth.users, to provide simple
authentication.

This type of authentication isn't exactly secure, using plaintext
passwords, and isn't scalable since it's currently in the config file,
but for a small handful of users on a private network it's sufficient.

Add a util to ServiceSpec to make it easy to authorise requests
properly, and authorise all requests in existing tests

Add tests to ensure no authorisation, or bad authorisation, gives 401