Fix logback config
logback-test.xml was blanked to prevent annoying log messages appearing
in the test output - that's no good as it sets the dev config to blank
too.

Instead, wrap the dev logback config in an if to ensure it's only used
if environment is development. Amend the main logback file to only apply
to production.