Move the shutdown handler to Reaper
It makes more sense to belong to the Reaper than to Main
Also had a stray TODO in the Main.shutdown code

Move the service actor definitions back into the main method and expose
the reaper instead

The reaper can be sent the Reaper.KillAll message to massacre its
targets. I narrowly resisted the urge to name it Massacre.